Minor issue
git-svn-id: http://hdt-it.googlecode.com/svn/trunk@366 ca5865f0-ea67-0c57-7e18-94a3701c88e0

+++ hdt-lib/src/hdt/BasicHDT.cpp
@@ -497,7 +497,7 @@ void BasicHDT::loadTriplesFromHDTs(const char** fileNames, size_t numFiles, cons
 
 	        TripleID tid;
 	        char str[100];
-	        uint64_t j = 0;
+	        long long int j = 0;
 	        while(it->hasNext()) {
 	        	TripleString *ts = it->next();
